requirements.txt
pandas
tldextract
dnspython
requests
beautifulsoup4
sentence-transformers
openpyxl